+++ src/adPEAS.ps1
@@ -189,6 +189,7 @@ if ($Script:ScriptPath) {
     . "$Script:ScriptPath\modules\Helpers\Get-PrincipalContainer.ps1"
     . "$Script:ScriptPath\modules\Helpers\Get-GPOEffectiveness.ps1"
     . "$Script:ScriptPath\modules\Helpers\ConvertTo-LDAPDNFilter.ps1"
+    . "$Script:ScriptPath\modules\Helpers\ConvertTo-UACValue.ps1"
     . "$Script:ScriptPath\modules\Helpers\Get-OUPermissions.ps1"
     . "$Script:ScriptPath\modules\Helpers\Get-GPOLinkage.ps1"
     . "$Script:ScriptPath\modules\Helpers\Test-adPEASLicense.ps1"